sign_in/requirement.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
flask-jwt-extended==4.6.0
redis==5.2.0
Werkzeug==3.1.2
